requirements.txt
flask==3.0.3
python-dotenv==1.0.1
flask_sqlalchemy==3.1.1
flask-marshmallow==1.2.1
flask-restful==0.3.10